Add comment about MTU 9000 setting
Red Hat Ceph documentation recommends MTU 9000 (see the end
of section 2.4 of [1]). Since enp7s0 is hosting Ceph traffic
this setting of 9000 is correct.

Add a comment next to the setting stating that in our case
there's one NIC, enp7s0 so we used that MTU. But if an
environment has >1 NIC and the storage traffic is only on
a particular NIC, then we could use a different MTU for
other NICs.

[1] https://access.redhat.com/documentation/en-us/red_hat_ceph_storage/6/html/hardware_guide/general-principles-for-selecting-hardware#network-considerations-for-red-hat-ceph-storage_hw

Signed-off-by: John Fulton <fulton@redhat.com>