Do not put empty services list in OpenStackDataPlaneNodeSet
If the services field is omitted from the spec in
OpenStackDataPlaneNodeSet, then the default services
list in the dataplane-operator will be created.

The problem with having a services list in NodeSet but
setting it to an empty list is that the services will
not be created. By ommiting the field we can get the
services created but also have a clear interface where
the user can define their services in the deployment
for pre and post ceph as originally introduced by
eaf0392b1bd1a52d0728be8d1aef54b4360dcd3c

Signed-off-by: John Fulton <fulton@redhat.com>