Restructure for kustomize
This patch solves the following problems by using kustomize:

A. Duplication: A new VA/DT would result in duplicate files
B. Substitution: User must update every line with a #CHANGEME

kustomize inheritance is insufficient to solve these problems
so this patch uses components [1]. The original CRs are moved
to a shared library directory with new names:

- stage1 to olm-deps
- stage2 to olm-osp
- stage3 to networking
- stage4 to control-plane

The duplication problem is solved by having each VA/DT use a
kustomization file which composes only the required CRs from
the common library. A common/ directory is also introduced for
CRs which all VAs/DTs are expected to need (e.g. OLM).

The substitution problem is solved by using `replacements`
(see `lib/control-plane/kustomization.yaml`) from a
user-provided file (see `examples/va/hci/values.yaml`).

The new examples/ directory becomes the user's main interface
and the following generates the CRs from stages 1-4:
```
kustomize build examples/common/olm/
kustomize build examples/va/hci/
```
In `examples/va/hci/kustomization.yaml` a URL is used to
reference a copy of this repository so that this repository
can be the upstream of a user's own kustomize pipeline.

Note: stage5/ and stage6/ are renamed to edpm-pre-ceph/ and
edpm-post-ceph/ respectively but are left in the user-facing
hci/ directory since they are not yet migrated to kustomize.
Instructions are updated for the new location but not yet in
how they are used.

This is based on work [2] originally proposed by Matthew Booth.

[1] https://github.com/kubernetes/enhancements/tree/master/keps/sig-cli/1802-kustomize-components
[2] https://github.com/openstack-k8s-operators/architecture/pull/28

Signed-off-by: John Fulton <fulton@redhat.com>