Add compute-starter-kit dt
This change adds a readme describing the compute starter kit
DT and job variants. This DT may be graducated to a VA in the future
and will be used as the basis of the compute content promotion jobs
prior to integration of promoted content.

This initial patch also include a defintion of the compute-kit
controlplane containing only keystone, placement, glance, neutron
and nova. it does not contain a dataplane definition.
that will be added in later patches.

this dt currently deploy nova with only cell 0
cell 1 will be added when adding the dataplane