Add missing links now that the SRIOV VA has merged
Add links to SRIOV VA from docs `mkdocs.yml` and top
level README.md.

Create a generic workarounds section in the docs
directory for a the NNCP workaround. Link to it
from both the HCI and SRIOV docs to deduplicate
content. Add an alternative method to resolve the
NNCP workaround.

Some links 404'd so the paths have been corrected.

Signed-off-by: John Fulton <fulton@redhat.com>