Do not deploy GlanceAPI pods by default
While we still apply the Glance CR during the first ctlplane deployment, this
patch sets both internal and external instances to replicas: 0.
By doing this we prevent a scenario where Glance is run without a supported
backend, causing issues as long as RWX is not supported and we still apply a
"split" layout by default.

We should split by default at this stage for at least two reasons:

1. a layout can't be change, so if "file" is used and a single instance
   is deployed, it's not possible to patch the "default" glance instance
   and change its layout (it should be entirely deleted and recreated);

2. in the next phase (stage5) we're going to patch the ctlplane and
   apply a backend (e.g., ceph)

It's still possible to kustomize at this stage the ctlplane and start
with a backend different than Ceph (e.g., swift).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>