Create MetalLB and NMState CRs in common (before NNCP CRs)
This patch moves the creation of the MetalLB and NMState
CRs to examples/common based on the assumption that all
VAs and DTs will need them. lib/deploy is divided into
lib/nmstate and lib/metallb. The common README is updated
with commands to confirm MetalLB is ready and NMState is
available. The HCI and NFV SRIOV VAs are updated accordingly.

Without this patch running kustomize in examples/va/hci
or examples/va/nfv/sriov results in a single CR containing
MetalLB, NMState and NNCP resources. When all of them are
created at the same time, OpenShift will try to create the
NNCP CRs while the NMState CRs are still installing and
produce a resource mappings not found error. Moving MetalLB
and NMState to common and updating the common README to add
a step to wait for them has a side effect of resolving this
problem.

The workaround section of the docs directory for this issue
has been removed though the workaround docs have been changed
into an FAQ so that the procedure of building smaller CRs
with separate components is still documented.

Jira: OSPRH-3743
Signed-off-by: John Fulton <fulton@redhat.com>