Uni* DTs: a bunch of proven, multicomponent set of deployments
The deployment part of this proposal have been used to combine
and test together a good amount of parts which "makes sense"
together.
In fact what is available now as VA1 was originally inspired
by one of these topologies (if it wouldn't have been created
as VA1, that would have been uni-gamma).