Enable Swift in VA2 and set it as Glance backend
Glance is set to replicas: 0 by default until a backend is chosen.
VA1 has ceph, and this means that the control plane is patched at
stage 5 to set it as a backend (and set replicas: 1).
In VA2, however, stage3 doesn't enable Glance or set a backend.
This patch does two things:
1. Enable Swift in VA2 (VA1 had RGW provided by Ceph, it makes
   sense to have Swift here);
2. Set Glance with the Swift backend and increase replicas to 1

By doing this we can test VA2 with a backend different than Ceph.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>