Use nic aliases in the edpm network template
To allow easier automation and avoid modifying the os-net-config
template of each VA to just change the interface name of the bridge
change the explicit interface name to an alias.
This requires to add the os-net-config interfaces alias table
to ensure proper translation between the alias and the interface
name.