Add automation directory
The automation directory contains YAML files which define
the stages for each VA/DT including the path to the CR
kustomization and its values as well as a shell command
which may be used to validate that the stage is complete.

These YAML files may be parsed by external tools which
apply the CRs in stages to automate the full deployment.

Because the structure of the architecture may change the
automation directory is separate and at the top of the
repository so that it does not need to change with the
rest of the architecture repo.

Signed-off-by: John Fulton <fulton@redhat.com>