Bump control plane and dataplane deployment timeout to 30m
Recently we have added VA HCI ci in downstream using
ci-framework devscript reproducer. Sometimes control plane
or dataplane deployment does not finishes with existing
timeout.

In order to finish the deployment, we are bumping the timeout
to 30 m for controlplane and dataplane deployment.

In upstream Zuul CI, the controlplane deploy timeout[1] is 30 mins
and EDPM deploy timeout[2] is 40 mins.

Links:
[1]. https://github.com/openstack-k8s-operators/ci-framework/blob/main/roles/edpm_prepare/defaults/main.yml#L27
[2]. https://github.com/openstack-k8s-operators/ci-framework/blob/main/roles/edpm_deploy/defaults/main.yml#L25

Signed-off-by: Chandan Kumar <chkumar@redhat.com>