Provide documentation for pull request testing
Provide documentation with an overview of testing expected to be
supplied along with pull requests in order to prove that changes are
operational, and do not cause environment regressions. Going forward,
authors of pull requests will be expected to provide manual testing of
the changes being proposed, so that we can attempt to avoid regressions
within the architecture repository.

Signed-off-by: Leif Madsen <lmadsen@redhat.com>