Move octavia-ca-passphrase to properly apply NamespaceTransformer
For some reason, when ci-framework generates the
octavia-ca-passphrase Secret, it does not include any namespace.
The output obtained with `kustomize build` is different and the
openstack namespace is correctly included in this case.

With this patch, the octavia-ca-passphrase secret is moved to the
kustomization.yaml file where the NamespaceTransformer is defined too.
After this change, the openstack namespace is properly added to the
secret.