[uni01alpha] Add AZ parameter for octavia and re-enable
The uni01alpha uses AZs, some of which are not available at the time
octavia is deployed. A recent patch to support using AZs in the octavia
deployment should allow octavia to proceed as soon as the ovn controller
in the control plane is available.

This patch adds the appropriate value for uni01alpha and re-enables
octavia.

Note: this also fixes an issue with a mismatch with what octavia expects
for the IP range for the network attachment. While this might be
configurable in the operator in the future, it will remain 172.24.x.x as
a default for IPv4.

Depends-On: https://github.com/openstack-k8s-operators/octavia-operator/pull/302
Jira: OSPRH-6901