Add PIDONE VA
This commit add a "PIDONE Validated Architecture" composed by:

- 3 x ocp masters
- 3 x ocp workers
- 1 x testOperator-tainted ocp worker

The additional workers are required so that we can perform
disruptive tests simulating ocp failures (reboots/crashes/etc).

Similarly, we need a special worker that would only run
testoperator pods so that the tests will not be affected by
the simulated infrastructure failures.