Cinder wait for DB writes on reads
Because we deploy the database in multi-master mode we can have cases
where a service writes something in the database and when another one
reads from the DB that data is not yet there.

This is a problem for Cinder, so we need to ensure that reads happen
after writes from other nodes are present.

The cinder operator already has this [1], but until jobs are using that
code we will be adding the configuration option manually.

[1]: https://github.com/openstack-k8s-operators/cinder-operator/commit/b0747944394807862e4cdcfa7052f1f8d1febf94