Set HAProxy timeout to 60 seconds for storage
This patch sets the HAProxy timeout to 60 seconds for storage services
(Cinder, Glance, and Manila).

This way it's in sync with the Apache default timeouts and matches what
we have in OSP17.

This will be fixed in the operators themselves to have these same
defaults, and this patch is compatible with that change.

When the operators code is merged and is run by the CI jobs the defaults
in the operators will be ignored in favor of the defaults defined in
this patch, so we'll want to revert this patch then.

Jira: https://issues.redhat.com/browse/OSPRH-7393
Jira: https://issues.redhat.com/browse/OSPRH-7415