Ensure certificates are installed before Ceph is deployed
In order for Ceph RGW to use HTTPS we need to make sure
the `install-certs` service runs on EDPM nodes which host
Ceph. This service must run before Ceph is configured by
the human operator or ci-framework. As per ci-framework
PR 1821 the cifmw_cephadm Ansible role will stat the files
installed by the `install-certs` service in order to know
if HTTP or HTTPS will be used for the RGW service that it
configures.

Add `install-certs` service to edpm-pre-ceph nodeset for
uni04delta (the non-HCI ceph job). Prior to this patch
only the edpm-nova nodeset was using the `install-certs`
service.

Add `install-certs` service to the edpm-pre-ceph run for
HCI VA. It should not be removed from the edpm-post-ceph
run because the dataplane goes through the list of enabled
services to determine which certs to create. There is no
harm in the `install-certs` service being called again.

Required for Jira: https://issues.redhat.com/browse/OSPRH-7245

Signed-off-by: John Fulton <fulton@redhat.com>