Unique pod names for cinder and glance
Cinder pods need to be uniquely name across all jobs that use the same
storage backend.

Glance also needs unique names when using Cinder as a backend.

From a CI point of view there is no downside to having unique names,
since all that changes are the CR and pod names for cinder and glance,
so we make the change permanent for all jobs.

Depends-On: https://github.com/openstack-k8s-operators/openstack-operator/pull/847
Depends-On: https://github.com/openstack-k8s-operators/cinder-operator/pull/400
Depends-On: https://github.com/openstack-k8s-operators/glance-operator/pull/555
Jira: OSPRH-7396