Epsilon: Re-enable multipath
Previously we disabled multipath in Epsilon jobs, but now that we are
using the webhook to change iSCSI login timeouts [1] they can be
re-enabled.

As a side note, and for future reference, disabling multipathing should
not have been in this architecture repo, because it's not an
architecture specific thing, it's a deployment specific issue.

The right place to change this would be in the job specific repo.

[1]: https://github.com/openstack-k8s-operators/ci-framework/pull/2060