Use pre/post Ceph servicesOverride in HCI deployments
Prior to this patch the OpenStackDataPlaneNodeSet changed
before and after Ceph was deployed. This resulted in an
incomplete nodeset services definition which resulted in
incomplete updates in future deployments.

With this change the NodeSet will have the full services
list but the pre-ceph and post-ceph deployments each pass
a different servicesOverride list so the desired subset
of services can be run before or after Ceph is deployed
but the NodeSet services list remains complete.

Note that the NodeSet does not have the `nova-custom-ceph`
service when it's initially created since that service
is not defined until after Ceph is deployed. However,
When `kustomize build` is run in `examples/va/hci`
the full services list is defined in the NodeSet.
Regardless, only a subset of services are run for
the pre and post Ceph deployments.

Jira: https://issues.redhat.com/browse/OSPRH-9172

Signed-off-by: John Fulton <fulton@redhat.com>