Add Swift network for RGW to HCI scenario
When Ceph RGW is used, an endpoint for Swift storage is
hosted not in a pod on k8s but on an EDPM node. Thus, a
service hosted on an EDPM node will need to be accessed
by cloud users from a separate network.

This patch adds the Swift storage network (swift) with
VLAN 25 and range 172.22.0.0/24 in the HCI values example.
The Swift network is configured on the HCI EDPM nodes
and an NNCP, NAD, L2Advertisement and IPAddressPool are
defined so that a pod in k8s can connect to it; such as
the tempest pod which will perform object storage tests.

In order to make these changes va/hci now keeps its own
copy of the nncp and networking directories since they
differ (by the new network) from the generic ones in the
lib directory.

Jira: https://issues.redhat.com/browse/OSPRH-6675
Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/2301

Signed-off-by: John Fulton <fulton@redhat.com>