Introduce "OSP" infra snippet in the automation file
For adoption, there's a need to override the infra exposed by the
scenario.
While this new "adoption" entry creates a stronger link with the
ci-framework, it can also be consumed by any other tool, as long as it
knows how to consume the two "patches".

This is part of a PoC, and will probably lead to discussions regarding
the best way to manage all of the datasets.