Add NFS network for Manila to uni04delta
Manila Tempest tests need to connect to the share
for Ganesha via a special (openstack) network [1].

This patch adds the NFS storage network with VLAN
24 and range 172.21.0.0/24 in uni04delta. The NFS
network is connected to Ceph and Compute EDPM nodes.
A NNCP, NAD, L2Advertisement and IPAddressPool are
defined for the NFS network so that a pod in k8s can
connect to it; such as the tempest pod which will
perform the storage tests and the manilaShares pod(s).

[1] https://opendev.org/openstack/manila-tempest-plugin/src/branch/master/manila_tempest_tests/config.py#L99

Jira: https://issues.redhat.com/browse/OSPRH-7417
Depends-On: openstack-k8s-operators/ci-framework#2273

Signed-off-by: John Fulton <fulton@redhat.com>