Bump edpm-deployment wait condition timout
This patch is in response to the linked Jira card where we can see
enabling FIPs in the uni07eta job causes the deployment to fail by a few
minutes. This is due to the increased time enabling FIPs adds caused by
the EDPM nodes requiring a reboot.

Jira: https://issues.redhat.com/browse/OSPRH-11095