Set Cinder customServiceConfig globally in DCN DT
The customServiceConfig for cinder in the DCN DT is used
to set the storage_availability_zone. We wish to do this
for all Cinder services in the default site (`az0`), not
just for cinderAPI or cinderBackup as the AZ was missing
from the cinder-scheduler. Thus, we will move it to the
global section under the cinder template and out of the
cinderAPI and cinderBackup since they will inherit it.
We still need to set the backend_availability_zone in
all cinderVolumes except az0 since it will also inherit
the new default value. However, the additional line is
kept in az0's customServiceConfig to make the pattern
easier to see for az0, az1, az2, etc.

Jira: https://issues.redhat.com/browse/OSPRH-11915

Signed-off-by: John Fulton <fulton@redhat.com>