BGP controlplane Pure L3 Spine&Leaf
With this PR, BGP DT01 configures isolated L2 segments for networkers and
computes.
The controlplane communications between RHOSO services running on OCP
nodes and services running on EDPM nodes is routed through the BGP
network. Spines and leaves connect the computes, networkers and OCP
nodes in different racks and route the traffic between them, as well as
the dataplane traffic to/from the Openstack VM instances.

OSPRH-13820