Decouple uni01alpha network stage from control plane stage
This removes the networking components...

    MetalLB
    NetConfig
    Net-attach-def

...from the control plane stage of dt/uni01alpha and gives them their
own separate stage in between the NNCP stage and the control plane
stage.

This is needed for the uni01alpha adoption job, which relies upon
this DT. For adoption we want to create the NNCPs, MetalLB CRs,
NetConfig CR and Net-attach-defs, but do NOT want to create the
OpenStackControlPlane.

Following this can this can be done by creating an automation file
with only two stages, one for NNCPs and the other for the other
networking components.

This change was created using the following PR as inspiration:
  https://github.com/openstack-k8s-operators/architecture/pull/489

Jira: OSPRH-13801