Add OpenStack OLM subscription configurations
Ability to install a specific version of the operator(s) when multiple
versions are available in the OLM catalog.

Added new kustomization files and YAML configurations for OpenStack
operator subscriptions using OLM. This includes catalog sources,
namespaces, and operator groups for openstack-operators, as well as
subscription definitions for the OpenStack components when the version
is below v1.0.7.

Kustomization facilitates version-specific configurations starting
with v1.0.3 and v1.0.6, incorporating modifications like the
elimination of the AnsibleEE operator in subsequent versions. From
version 1.0.7 onwards, only the OpenStack operator is retained.

For each overlay we define the same set of ConfigMap that reach a set
of relevant parameter in the subscriptions and catalog source
definitions.

An unexpected mechanism here is that we override the
`common/olm-subscriptions/kustomization.yaml` file from ci-framework
for version `v1.0.3` and `v1.0.6`.  This enables the dynamic selection
of the right overlay.  For version `v1.0.7` and beyond we used the
default overlay and we don't need to overwrite the `kustomization`
file.

The problems that we cannot solve with just parameters are:
- `v1.0.3` and `v1.0.6` have a different set of subscriptions to
include
- `startingCSV` parameter is "dynamic" as it's composed of the
`spec.name`.`version`, not just `version`. Thus we cannot encode in a
kustomize parameter without create our own transformer. As
transformwer are not deployed here, it would be too much work to
boostrap it.

All this disapear starting with `v1.0.7` where we have only one
subscription to set.