Add namespace replacement for openstack-catalog and subscription
Only for the olm-subscription part.

In the `lib/olm-openstack/subscription.yaml` file, used for most
deployments, the `CatalogSource` is using the `openstack-operators`
namespace.

In the `lib/olm-openstack-subscriptions/base/catalogsource.yaml`, the
`CatalogSource` is using the cluster-wide `openshift-marketplace`
namespace. This configuration is used for running update CI tests.

Both configurations work, but some CI tools are hardcoded to have the
`CatalogSource` set in the `openstack-operators` namespace.

To accommodate the CI tool, we add a new variable
`openstack-catalog-namespace` that can adjust the `CatalogSource`
namespace for the `olm-openstack-subscription`.

We switch the default to `openstack-operators` so all jobs are in line.

Closes: [OSPRH-18002](https://issues.redhat.com/browse/OSPRH-18002)