[fr4] Adjust nova04delta T-O, add DNS and BMO layers
1) Add DNS config and ctlplane NAD for T-O

When T-O uses ctlplane net for metrics scrapping on EDPM hosts,
additional Openshift DNS servers must be added to let it resolving
FQDN of EDPM hosts on controlplane network. Note that cifmw addresses
this differently, however, nova04delta is also used outside of cifmw.

Enable metricStorage/dashboard.
Also, as documented for T-O, add ctlplane NAD for metrics storage
template spec in oscp CR.

2) Simplify bmh kustzomiation and extract BMO into a separate
kustomization for allowing it in parallel steps in CI.

Create a separate kustomization for BMH and provisioning to make it
applicable independently of nodeset.

Adjust automation to create CP and BMH in parallel.

The stages will now execute as follows:
1. nncp-configuration (networking setup)
2. network-configuration (MetalLB setup)
3. control-plane and baremetalhosts-configuration (run in parallel)
4. edpm-nodeset (depends on both control plane and BMH being ready)
5. edpm-deployment (final deployment)

3) Align nvidia-vfio VA with underlying DT changes

4) nova04/vfio: fix storage class overriding

The storage lib component relies on network-values data for setting
control plane SC values, historically. The replacement of telemetry
data overrides this in the nova04delta and vfio DT/VA.

Fixing that by re-applying the storage component logic locally in the
DT/VA as well, while repurposing it to control-plane data instead of
network-values didn't work in CI (however, it works locally with
kustomize build).

Fix that by simply providing pvcStorageClass directly in DT/VA.

We would not have to fix that, if T-O had been taking the top scope
storageClass for its metricStorage template (that fix will be in FR5
likely).

Unrelated change:
* update gitignore for asdf .tool-versions to simplify local
development when kustomize is installed with asdf.

CI impact: rework extracted BMO values for baremetalhosts-configuration

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>