nova04/vfio: fix storage class overriding
The storage lib component relies on network-values data for setting
control plane SC values, historically. The replacement of telemetry
data overrides this in the nova04delta and vfio DT/VA.

Fixing that by re-applying the storage component logic locally in the
DT/VA as well, while repurposing it to control-plane data instead of
network-values didn't work in CI (however, it works locally with
kustomize build).

Fix that by simply providing pvcStorageClass directly in DT/VA.

We would not have to fix that, if T-O had been taking the top scope
storageClass for its metricStorage template (that fix will be in FR5
likely).

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>